continued refactoring code
