+++ functions/tests/lyft.test.js
@@ -2,6 +2,7 @@ const axios = require("axios");
 jest.mock("axios");
 const { getLyftPrices } = require("../middleware/lyft");
 const mockConfig = require("firebase-functions-test")();
+const { functions, responseBody, params, response } = require("./constants");
 
 mockConfig.mockConfig({
   getprices: {
@@ -9,41 +10,6 @@ mockConfig.mockConfig({
     lyft_endpoint: "www.test1.com",
   },
 });
-
-const functions = {
-  config: () => {
-    return {
-      getprices: {
-        uber_endpoint: "www.test.com",
-        lyft_endpoint: "www.test1.com",
-      },
-    };
-  },
-};
-
-const response = {
-  status: (status) => {
-    response.statusCode = status;
-    return response;
-  },
-  send: ({ error, status, message }) => {
-    response.body = {
-      error: error,
-      status: status,
-      message: message,
-    };
-    return response;
-  },
-};
-
-const params = {
-  startingLatitude: 123,
-  startingLongitude: 123,
-  endLongitude: 123,
-  endLatitude: 123,
-};
-
-const responseBody = {};
 
 describe("getLyfPrices Testing Suite", () => {
   beforeEach(() => {
